@@ -1,24 +1,32 @@
 import {inject} from 'aurelia-framework';
-import ImageInfo from '../model/image_info';
+import AppContext from '../app/context';
 import {EVENTS, EventSubscriber} from '../events/events';
-import {customElement} from 'aurelia-framework';
+import {customElement, bindable} from 'aurelia-framework';
 import {slider} from 'jquery-ui';
 
 @customElement('custom-dimension-slider')
-@inject(ImageInfo, Element)
+@inject(AppContext, Element)
 export default class CustomDimensionSlider extends EventSubscriber {
-    sub_list = [
-        [EVENTS.FORCE_UPDATE, (image_id) => this.forceUpdate(image_id)],
-        [EVENTS.FORCE_CLEAR, () => this.detached()]
-    ];
-    constructor(image_info, element) {
-        super(image_info.eventbus);
-        this.subscribe();
+    image_info = null;
+    @bindable dim = 't';
+    sub_list = [[EVENTS.FORCE_CLEAR, (params = {}) => {
+        this.image_info = null; this.forceUpdate() }]];
+    sub_list = [[EVENTS.FORCE_UPDATE, (params = {}) => {
+        this.image_info = this.context.getSelectedImageConfig().image_info;
+        this.forceUpdate() }]];
 
-        this.image_info = image_info;
+    constructor(context, element) {
+        super(context.eventbus);
+        this.context = context;
         this.element = element;
-        this.dim = this.element.id === 'time' ? 't' : 'z';
-        this.elSelector = "#" + this.element.id + " [name='dim']";
+        this.elSelector = null;
+    }
+
+    bind() {
+        this.elSelector =
+            "[au-target-id=" + $(this.element).attr("au-target-id") +
+            "] div[name='dim']";
+        this.subscribe();
     }
 
     attached() {
@@ -29,32 +37,44 @@ export default class CustomDimensionSlider extends EventSubscriber {
         try {
             $(this.elSelector).slider( "destroy" );
         } catch (ignored) {}
-        $(this.elSelector).hide();
+        this.hide();
+    }
+
+    show() {
+        $(this.element).css('visibility', 'visible');
+        //$(this.element).show();
+    }
+
+    hide() {
+        //$(this.element).hide();
+        $(this.element).css('visibility', 'hidden');
     }
 
     onChange(value) {
         if (value === this.image_info.dimensions[this.dim]) return;
 
         this.image_info.dimensions[this.dim] = parseInt(value);
-        this.image_info.eventbus.publish(
+        this.context.publish(
             EVENTS.DIMENSION_CHANGE,
             {dim: this.dim, value: [this.image_info.dimensions[this.dim]]});
     }
 
     forceUpdate() {
         this.detached();
+
         if (this.image_info.dimensions['max_' + this.dim] <= 1) return;
 
         $(this.elSelector).slider({
             orientation: this.dim === 'z' ? "vertical" : "horizontal",
-            min: 0, max: this.image_info.dimensions['max_' + this.dim],
+            min: 0, max: this.image_info.dimensions['max_' + this.dim] - 1 ,
             step: 1, value: this.image_info.dimensions[this.dim],
             change: (event, ui) => this.onChange(ui.value)
         });
-        $(this.elSelector).show();
+        this.show();
     }
 
     unbind() {
-        this.unsubscribe();
+        this.unsubscribe()
+        this.image_info = null;
     }
 }
