@@ -1,48 +1,39 @@
-import {Container} from 'aurelia-dependency-injection';
-import {BindingEngine, noView} from 'aurelia-framework';
-import {EVENTS} from '../events/events.js';
+import {noView} from 'aurelia-framework';
+import {EVENTS} from '../events/events';
+import Misc from '../utils/misc';
 
 @noView
 export default class ImageInfo {
-    data_received = false;
     dimensions = {t: 0, max_t : 1,z: 0, max_z : 1};
     channels = null;
     tiled = false;
     show_regions = false;
 
-    constructor(eventbus, server, initial_image_id) {
-        this.eventbus = eventbus;
-        this.binding = new Container().get(BindingEngine);
-        this.server = server;
-        this.image_id = initial_image_id;
+    constructor(context, image_id) {
+        this.context = context;
+        this.image_id = image_id;
         this.dataset_id = null;
-        this.imageIdObserver = null;
+    }
 
+    bind() {
         this.requestData();
-        this.registerImageIdObserver();
     }
 
-    unregisterImageIdObserver() {
-        if (this.imageIdObserver === null) return;
-
-        this.imageIdObserver();
-        this.imageIdObserver = null;
-    }
-
-    registerImageIdObserver() {
-        this.unregisterImageIdObserver();
-        this.imageIdObserver =
-            this.binding.propertyObserver(this, 'image_id').
-            subscribe(() => this.requestData(true));
+    unbind() {
+        this.dimensions = {t: 0, max_t : 1,z: 0, max_z : 1};
+        this.channels = null;
+        this.tiled = false;
+        this.show_regions = false;
     }
 
     requestData() {
-        this.data_received = false;
-        this.eventbus.publish(EVENTS.IMAGE_CHANGE, this.image_id);
-        var url = this.server + "/webgateway/imgData/" + this.image_id + '/';
+        let dataType = "json";
+        if (Misc.useJsonp(this.context.server)) dataType += "p";
+        let url = this.context.server + "/webgateway/imgData/" + this.image_id + '/';
         $.ajax(
             {url : url,
-            dataType : "jsonp",
+            dataType : dataType,
+            cache : false,
             success : (response) => {
                 if (typeof response.meta === 'object' &&
                         typeof response.meta.datasetId === 'number')
@@ -52,15 +43,12 @@ export default class ImageInfo {
                     t: 0, max_t : response.size.t,
                     z: 0, max_z : response.size.z
                 };
-                this.show_regions = false;
-                this.data_received = true;
-                this.eventbus.publish(EVENTS.FORCE_UPDATE, this.image_id);
-                if (this.dataset_id)
-                    this.eventbus.publish(EVENTS.INIT_THUMBSLIDER, this.dataset_id);
+                this.context.publish(
+                    EVENTS.FORCE_UPDATE,
+                        {dataset_id: this.dataset_id, image_id: this.image_id});
             },
             error : (error) => {
-                this.show_regions = false;
-                this.eventbus.publish(EVENTS.FORCE_CLEAR);
+                this.context.publish(EVENTS.FORCE_CLEAR);
             }
         });
     }
