@@ -1,7 +1,6 @@
-import ImageInfo from '../model/image_info';
-import RegionsInfo from '../model/regions_info';
+import AppContext from '../app/context';
 import {EVENTS, EventSubscriber} from '../events/events';
-import {inject, customElement} from 'aurelia-framework';
+import {inject, noView, customElement, bindable} from 'aurelia-framework';
 
 import DragAndSizable from '../controls/dragandsizable';
 
@@ -9,17 +8,19 @@ require('../css/viewer.css');
 import {ol3} from '../../libs/ome-viewer-1.0.js';
 
 @customElement('custom-viewer')
-@inject(ImageInfo, RegionsInfo)
+@inject(AppContext, Element)
 export default class CustomViewer extends EventSubscriber {
+    image_config = null;
+    dragizable = null;
     sub_list = [
         [EVENTS.IMAGE_CHANGE,
             (imageid) => {
-                this.reset();
-                this.viewer.changeToImage(imageid)
+                this.image_config = this.context.getSelectedImageConfig()
+                if (this.context.useMDI) this.reset();
+                this.viewer.changeToImage(imageid);
+                if (this.context.useMDI) this.viewer.redraw();
             }],
         [EVENTS.FORCE_UPDATE, () => this.forceUpdate()],
-        [EVENTS.FORCE_CLEAR,
-            () => console.info(this.image_info.data)],
         [EVENTS.SHOW_REGIONS, (flag) => this.showRegions(flag)],
         [EVENTS.MODIFY_REGIONS,
             (params={}) => this.viewer.modifyRegionsStyle(
@@ -30,45 +31,52 @@ export default class CustomViewer extends EventSubscriber {
         [EVENTS.REGIONS_VISIBILITY,
             (params={}) => this.updateRegionsVisibility(params.flag, params.rois)],
         [EVENTS.DRAW_SHAPE,
-            (type) => this.viewer.drawShape(type)],
+            (type) => {
+                this.viewer.redraw();
+                this.viewer.drawShape(type)
+            }],
         [EVENTS.DIMENSION_CHANGE, (data, event) =>
             this.viewer.setDimensionIndex.apply(
                 this.viewer, [data.dim].concat(data.value))],
         [EVENTS.VIEWER_RESIZE, () => this.viewer.redraw()]
     ];
-    constructor(image_info, regions_info) {
-        super(image_info.eventbus)
-        this.subscribe();
-        this.image_info = image_info;
-        this.regions_info = regions_info;
-        this.dragizable = new DragAndSizable(this.image_info.eventbus);
-
-        this.viewer =
-            new ol3.Viewer(this.image_info.image_id,
-                { eventbus : this.image_info.eventbus,
-                  server : this.image_info.server
-                });
-        this.showRegions(this.image_info.show_regions);
+    constructor(context, element) {
+        super(context.eventbus)
+        this.context = context;
+        this.element = element;
+        this.aurelia_id = $(this.element).attr("au-target-id");
+        this.container = 'ol3_viewer_' + this.aurelia_id;
+        this.image_config = this.context.getSelectedImageConfig();
     }
 
     attached() {
-        this.dragizable.createDraggable("custom-viewer");
-        this.dragizable.createResizable("custom-viewer");
+        this.viewer =
+            new ol3.Viewer(this.image_config.image_info.image_id,
+                { eventbus : this.context.eventbus,
+                  server : this.context.server,
+                  container: this.container
+                });
+        this.showRegions(false);
+        this.subscribe();
+        if (this.context.useMDI) {
+            this.dragizable =
+                new DragAndSizable
+                    (this.context.eventbus, this.aurelia_id);
+                this.dragizable.createDraggable();
+                this.dragizable.createResizable();
+        }
     }
 
     detached() {
-        this.dragizable.destroyDraggable();
-        this.dragizable.destroyResizable();
-    }
-
-    unsubscribe() {
-        if (this.subscriptions.length === 0) return;
-        while (this.subscriptions.length > 0)
-            (this.subscriptions.pop())();
+        if (this.dragizable) {
+            this.dragizable.destroyDraggable();
+            this.dragizable.destroyResizable();
+            this.dragizable = null;
+        }
     }
 
     updateRegionsVisibility(flag = false, rois = []) {
-        if (!this.image_info.show_regions ||
+        if (!this.image_config.image_info.show_regions ||
             rois.length === 0) return;
 
         this.viewer.setRegionsVisibility(flag, rois);
@@ -86,8 +94,9 @@ export default class CustomViewer extends EventSubscriber {
             this.viewer.enableRegionsContextMenu(false);
             this.viewer.setRegionsModes([ol3.REGIONS_MODE.DEFAULT]);
             this.viewer.setRegionsVisibility(false, []);
-            for (var [key, value] of this.regions_info.data)
-                this.regions_info.data.get(key).selected = false;
+            if (this.image_config.region_info)
+                for (var [key, value] of this.image_config.region_info.data)
+                    this.image_config.region_info.data.get(key).selected = false;
         }
     }
 
@@ -97,15 +106,16 @@ export default class CustomViewer extends EventSubscriber {
             this.viewer.destroyViewer();
             this.viewer = null;
         }
+        this.image_config = null;
     }
 
     forceUpdate()  {
-        this.showRegions(this.image_info.show_regions);
+        this.showRegions(this.image_config.image_info.show_regions);
 
         var presentZ = this.viewer.getDimensionIndex('z');
         var presentT = this.viewer.getDimensionIndex('t');
-        var newZ = this.image_info.dimensions.z;
-        var newT = this.image_info.dimensions.t;
+        var newZ = this.image_config.image_info.dimensions.z;
+        var newT = this.image_config.image_info.dimensions.t;
 
         if (presentZ !== newZ)
             this.viewer.setDimensionIndex.apply(
@@ -117,11 +127,14 @@ export default class CustomViewer extends EventSubscriber {
     }
 
     reset() {
-        this.dragizable.resetPosition();
-        this.dragizable.resetSize();
+        if (this.dragizable) {
+            this.dragizable.resetPosition();
+            this.dragizable.resetSize();
+        }
     }
 
     closeMe() {
-        //TODO: kill widget
+        this.unbind();
+        $("custom-viewer").remove();
     }
 }
